Upgrade with seed refactoring
