+++ core/src/main/java/org/seedstack/audit/internal/AuditPlugin.java
@@ -13,7 +13,7 @@ import io.nuun.kernel.api.plugin.context.InitContext;
 import io.nuun.kernel.api.plugin.request.ClasspathScanRequest;
 import io.nuun.kernel.core.AbstractPlugin;
 import org.apache.commons.configuration.Configuration;
-import org.seedstack.audit.api.TrailExceptionHandler;
+import org.seedstack.audit.TrailExceptionHandler;
 import org.seedstack.audit.spi.TrailWriter;
 import org.seedstack.seed.core.internal.application.ApplicationPlugin;
