+++ core/src/main/java/org/seedstack/audit/internal/AuditedInterceptor.java
@@ -16,13 +16,13 @@ import net.jodah.typetools.TypeResolver;
 import org.aopalliance.intercept.MethodInterceptor;
 import org.aopalliance.intercept.MethodInvocation;
 import org.apache.commons.lang.StringUtils;
-import org.seedstack.audit.api.AuditService;
-import org.seedstack.audit.api.Trail;
-import org.seedstack.audit.api.TrailExceptionHandler;
-import org.seedstack.audit.api.Audited;
-import org.seedstack.seed.el.api.ELContextBuilder;
-import org.seedstack.seed.el.api.ELContextBuilder.ELPropertyProvider;
-import org.seedstack.seed.el.api.ELService;
+import org.seedstack.audit.AuditService;
+import org.seedstack.audit.Trail;
+import org.seedstack.audit.TrailExceptionHandler;
+import org.seedstack.audit.Audited;
+import org.seedstack.seed.el.ELContextBuilder;
+import org.seedstack.seed.el.ELContextBuilder.ELPropertyProvider;
+import org.seedstack.seed.el.ELService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
