+++ core/src/test/java/org/seedstack/audit/api/AuditEventTest.java
@@ -8,14 +8,14 @@
 /*
  * Creation : 23 juil. 2014
  */
-package org.seedstack.audit.api;
+package org.seedstack.audit;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 
 import org.junit.Test;
-import org.seedstack.audit.api.AuditEvent;
-import org.seedstack.audit.api.Trail;
+import org.seedstack.audit.AuditEvent;
+import org.seedstack.audit.Trail;
 
 public class AuditEventTest {
