+++ core/src/test/java/org/seedstack/audit/internal/AuditConfigurerTest.java
@@ -23,8 +23,8 @@ import java.util.Set;
 import org.apache.commons.configuration.Configuration;
 import org.junit.Before;
 import org.junit.Test;
-import org.seedstack.audit.api.AuditEvent;
-import org.seedstack.audit.api.TrailExceptionHandler;
+import org.seedstack.audit.AuditEvent;
+import org.seedstack.audit.TrailExceptionHandler;
 import org.seedstack.audit.internal.AuditConfigurer;
 import org.seedstack.audit.internal.SecurityTrailExceptionHandler;
 import org.seedstack.audit.spi.TrailWriter;
