+++ core/src/test/java/org/seedstack/audit/internal/AuditContextTest.java
@@ -14,7 +14,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 
 import org.junit.Test;
-import org.seedstack.audit.api.Trail;
+import org.seedstack.audit.Trail;
 import org.seedstack.audit.internal.AuditContext;
 
 public class AuditContextTest {
