+++ core/src/test/java/org/seedstack/audit/internal/AuditModuleTest.java
@@ -20,8 +20,8 @@ import java.util.Set;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.internal.util.reflection.Whitebox;
-import org.seedstack.audit.api.AuditEvent;
-import org.seedstack.audit.api.TrailExceptionHandler;
+import org.seedstack.audit.AuditEvent;
+import org.seedstack.audit.TrailExceptionHandler;
 import org.seedstack.audit.internal.AuditConfigurer;
 import org.seedstack.audit.internal.AuditModule;
 import org.seedstack.audit.internal.SecurityTrailExceptionHandler;
