+++ core/src/test/java/org/seedstack/audit/internal/AuditedInterceptorTest.java
@@ -23,13 +23,13 @@ import org.aopalliance.intercept.MethodInvocation;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.internal.util.reflection.Whitebox;
-import org.seedstack.audit.api.AuditService;
-import org.seedstack.audit.api.TrailExceptionHandler;
-import org.seedstack.audit.api.Audited;
+import org.seedstack.audit.AuditService;
+import org.seedstack.audit.TrailExceptionHandler;
+import org.seedstack.audit.Audited;
 import org.seedstack.audit.internal.AuditedInterceptor;
-import org.seedstack.seed.el.api.ELContextBuilder;
-import org.seedstack.seed.el.api.ELContextBuilder.ELPropertyProvider;
-import org.seedstack.seed.el.api.ELService;
+import org.seedstack.seed.el.ELContextBuilder;
+import org.seedstack.seed.el.ELContextBuilder.ELPropertyProvider;
+import org.seedstack.seed.el.ELService;
 
 public class AuditedInterceptorTest {
