+++ core/src/test/java/org/seedstack/audit/internal/DefaultAuditServiceTest.java
@@ -22,14 +22,14 @@ import java.util.Set;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.internal.util.reflection.Whitebox;
-import org.seedstack.audit.api.AuditEvent;
-import org.seedstack.audit.api.Host;
-import org.seedstack.audit.api.Trail;
+import org.seedstack.audit.AuditEvent;
+import org.seedstack.audit.Host;
+import org.seedstack.audit.Trail;
 import org.seedstack.audit.internal.DefaultAuditService;
 import org.seedstack.audit.spi.TrailWriter;
-import org.seedstack.seed.core.api.Application;
-import org.seedstack.seed.security.api.SecuritySupport;
-import org.seedstack.seed.security.api.principals.Principals;
+import org.seedstack.seed.Application;
+import org.seedstack.seed.security.SecuritySupport;
+import org.seedstack.seed.security.principals.Principals;
 
 public class DefaultAuditServiceTest {
