+++ logback/src/it/java/org/seedstack/business/audit/IntegrationTest.java
@@ -14,10 +14,10 @@ import javax.inject.Inject;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.seedstack.audit.api.Audited;
+import org.seedstack.audit.Audited;
 import org.seedstack.seed.it.SeedITRunner;
-import org.seedstack.seed.it.api.ITBind;
-import org.seedstack.seed.security.api.WithUser;
+import org.seedstack.seed.it.ITBind;
+import org.seedstack.seed.security.WithUser;
 
 @RunWith(SeedITRunner.class)
 public class IntegrationTest {
