+++ logback/src/main/java/org/seedstack/audit/infrastructure/logback/AuditLogbackLayout.java
@@ -10,10 +10,10 @@ package org.seedstack.audit.infrastructure.logback;
 import javax.el.ELContext;
 import javax.inject.Inject;
 
-import org.seedstack.audit.api.AuditEvent;
-import org.seedstack.seed.core.api.Configuration;
-import org.seedstack.seed.el.api.ELContextBuilder;
-import org.seedstack.seed.el.api.ELService;
+import org.seedstack.audit.AuditEvent;
+import org.seedstack.seed.Configuration;
+import org.seedstack.seed.el.ELContextBuilder;
+import org.seedstack.seed.el.ELService;
 
 import ch.qos.logback.classic.spi.ILoggingEvent;
 import ch.qos.logback.core.CoreConstants;
