+++ specs/src/main/java/org/seedstack/audit/api/AuditEvent.java
@@ -5,7 +5,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.seedstack.audit.api;
+package org.seedstack.audit;
 
 import java.text.SimpleDateFormat;
 import java.util.Date;
